Update dgambin.R
Gj can be calculated only once for all values 1:100, because pgamma allows a whole vector of quantiles as input. It is not necessary to be calculated for each j for each k for each octave. Taking the calculation of Gj out of the Pk function makes the Pk function to be more concise (the alpha parameter is not necessary any more within Pk). Note that when Gj is expressed as a vector, the whole calculation of s can be done over vectors instead of over single scalars. So "j" can be substituted by "1:100", and the for loop is therefore not necessary. The new version of this Pk function is >10 times faster than the old one, so the whole function dgambin improves its performance.